refactor: align log download with issue-attachment logs.html build
Log section download now goes through buildCaptureFiles (video/image embed
+ report) identical to the issue-tracker attachment. Extract buildCtx and
the capture-files input into buildEditorCapture as a single source shared
by IssueCreateModal (submit) and the panel log downloads, preventing drift.

- buildEditorCapture: buildEditorMarkdownContext / buildEditorLogsCaptureInput
  / downloadEditorLogsHtml
- IssueCreateModal delegates to the single source (userAttachments only on
  the submit path); drop now-unused selectors/imports
- downloadCapture: export triggerDownload, drop media-less downloadLogsHtml
